Dump Meizu Pro 5 Android sources
Imported from:
https://github.com/meizuosc/m576
commit 2d3fa1ca1fbc8b52c75b2bcc3784b7e52085768b ("first commit")
Imported from author: Meizu OpenSource <patchwork@meizu.com>

+++ drivers/leds/leds-lp55xx-common.h
@@ -108,7 +108,7 @@ struct lp55xx_led {
 	u8 led_current;
 	u8 max_current;
 	struct work_struct brightness_work;
-	u8 brightness;
+	int brightness;
 	struct lp55xx_chip *chip;
 };
